+++ css/csstools-design-tokens/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+### 1.2.0
+
+- Added support for design token file imports from npm packages.
+
+npm package : `@your-org/your-tokens`
+
+```json
+{
+	"font-size": 0.2
+}
+```
+
+```css
+@design-tokens url(node_modules://@your-org/your-tokens/tokens.json) format('style-dictionary3');
+
+.example {
+	font-size: design-token('font-size');
+}
+```
+
 ## 1.1.0
 
 - Added `designTokens.valueFunctionName` option to control the design-token function name.